+++ scripts/clean_up.sh
@@ -1,5 +1,6 @@
 find . -name "*.pyc" -exec rm -rf {} \;
 find . -name "*.py,cover" -exec rm -rf {} \;
+find . -name "*.orig" -exec rm -rf {} \;
 find . -name "__pycache__" -exec rm -rf {} \;
 rm -rf build/
 rm -rf dist/