+++ examples/simple/bar/forms.py
@@ -1,0 +1,6 @@
+from django import forms
+
+class MyForm(forms.Form):
+    """Test form."""
+
+    number = forms.IntegerField(max_value=200)