+++ examples/simple/runserver/bootstrap3-theme-django-1-11.sh
@@ -1,0 +1,1 @@
+./manage.py runserver 0.0.0.0:8000 --traceback -v 3 --settings=settings.bootstrap3_theme_django_1_11 --traceback -v 3