+++ examples/simple/urls.py
@@ -57,6 +57,9 @@ url_patterns_args = [
     # foo URLs:
     url(r'^foo/', include('foo.urls')),
 
+    # bar URLs:
+    # url(r'^bar/', include('bar.urls')),
+
     url(r'^$', TemplateView.as_view(template_name=fobi_home_template)),
 
     # django-fobi public forms contrib app: